+++ CHANGELOG.md
@@ -23,6 +23,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-4043 Display different warmup tooltip for Dexcom G7 sensors
 - YLP-4046 Improve the display of the parameter changes tooltip on Daily view
 - YLP-4047 Use smart rounding for TIR percentage values
+- YLP-4104 Removing trends tab from patient view in the mobile version of the website interface
 
 ### Engineering
 - YLP-3864 SBOM generation in pipeline